% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_issues}
\alias{fetch_issues}
\title{Fetch all jese4sci issues}
\usage{
fetch_issues()
}
\value{
\verb{[data.frame]} with columns:
\itemize{
\item github_username
\item track
\item lesson
\item url
\item is_open
}
}
\description{
Fetch all jese4sci issues
}
