% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_track_issues}
\alias{fetch_track_issues}
\title{Fetch jese4sci issues for one track}
\usage{
fetch_track_issues(
  repo = c("jese4sci-RC", "jese4sci-MOD", "jese4sci-DOC", "jese4sci-VAL")
)
}
\arguments{
\item{repo}{\verb{[character(1)]} Which jese4sci track repo to fetch issues from.}
}
\value{
\verb{[data.frame]} See \code{fetch_issues()}.
}
\description{
Fetch jese4sci issues for one track
}
\keyword{internal}
