% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{is_assignment}
\alias{is_assignment}
\title{Check if issue is an assignment}
\usage{
is_assignment(issue_title, repo)
}
\arguments{
\item{issue_title}{\verb{[character]} Vector ofGitHub issue titles. Must all be
from the same repository.}

\item{repo}{\verb{[character(1)]} Repository name e.g. \code{"jese4sci-RC"}}
}
\value{
\verb{[logical]} Same length as \code{issue_title}, indicating is issue
corresponds to an assignment or not.
}
\description{
Check if issue is an assignment
}
\keyword{internal}
